ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.2
sentencepiece==0.1.99
protobuf==4.25.1
scipy==1.11.4
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.4.0
pytesseract==0.3.13

# Headless browser fetch (optional)
selenium==4.24.0
undetected-chromedriver==3.5.5

# Gemini SDK
google-generativeai>=0.7.0

